// src/models/TypeWithdrawal.ts
export interface TypeWithdrawal {
  IdTypeWithdrawal: number;
  nameTypeWithdrawal: string | null;
  codeTypeWithdrawal: string | null;
}

export interface TypeWithdrawalCreate {
  nameTypeWithdrawal: string | null;
  codeTypeWithdrawal: string | null;
}

export interface TypeWithdrawalUpdate {
  nameTypeWithdrawal: string | null;
  codeTypeWithdrawal: string | null;
}